Add axis repositioning across x, y, and, z axises
Axis Repositioning is activated by offhand gripping and holding
the trigger (c_repositionButtonID) and saved when released.
Once in repositioning mode and holding trigger, the axis is cycled
by clicking on the A button (c_defaultPositionButtonID). The distance
on the axis is tied to the offhand button moves. Currently, the offhand
moves only on the X-axis, but the weapon will preview the final location

+++ Skeleton.cpp
@@ -2726,18 +2726,30 @@ namespace F4VRBody
 							_hasLetGoRepositionButton = false;
 							_repositionButtonHoldStart = duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count();
 							_startFingerBonePos = rt->transforms[boneTreeMap[offHandBone]].world.pos - _curPos;
-							_MESSAGE("Reposition Button Hold start: weapon %s", weapname);
+							_MESSAGE("Reposition Button Hold start: weapon %s mode: %d", weapname, _repositionMode);
 						}
 					}
 					else{
+						if (!_repositionModeSwitched && reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_defaultPositionButtonID)) {
+							_repositionMode = static_cast<repositionMode>((_repositionMode + 1) % (repositionMode::total + 1));
+							vrhook->StartHaptics(c_leftHandedMode ? 0 : 1, 0.1 * (_repositionMode + 1), 0.3);
+							_repositionModeSwitched = true;
+							_MESSAGE("Reposition Mode Switch: weapon %s %d ms mode: %d", weapname, _pressLength, _repositionMode);
+						}
+						else if (_repositionModeSwitched && !(reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_defaultPositionButtonID))) {
+							_repositionModeSwitched = false;
+						}
 						_pressLength = duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count() - _repositionButtonHoldStart;
-						if (reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_repositionButtonID) && _pressLength > c_holdDelay) {
-							vrhook->StartHaptics(c_leftHandedMode? 0 : 1, 0.05, 0.3);
-						} else if (!(reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_repositionButtonID))) {
+						if (!_repositionHapticFired && reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_repositionButtonID) && _pressLength > c_holdDelay) {
+							vrhook->StartHaptics(c_leftHandedMode? 0 : 1, 0.1 * (_repositionMode + 1), 0.3);
+							_repositionHapticFired = true;
+						}
+						else if (!(reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_repositionButtonID))) {
 							_repositionButtonHolding = false;
 							_hasLetGoRepositionButton = true;
+							_repositionHapticFired = false;
 							_endFingerBonePos = rt->transforms[boneTreeMap[offHandBone]].world.pos - _curPos;
-							_MESSAGE("Reposition Button Hold stop: weapon %s %d ms", weapname, _pressLength);
+							_MESSAGE("Reposition Button Hold stop: weapon %s %d ms mode: %d", weapname, _pressLength, _repositionMode);
 						}
 					}
 
@@ -2762,11 +2774,23 @@ namespace F4VRBody
 						fingerBonePos = rt->transforms[boneTreeMap[offHandBone]].world.pos;
 						bodyPos = _curPos;
 						if (_repositionButtonHolding) {
+							// this is for a preview of the move. The preview happens one frame before we detect the release so must be processed separately.
 							auto end = fingerBonePos - _curPos;
 							auto change = vec3_len(end) > vec3_len(_startFingerBonePos)? vec3_len(end - _startFingerBonePos) : -vec3_len(end - _startFingerBonePos);
 							if (change != 0) {
-								weap->m_localTransform.pos.x += change;
-								_DMESSAGE("Updating x translation for %s by %f to %f", weapname, change, weap->m_localTransform.pos.x);
+								switch (_repositionMode) {
+								case x:
+									weap->m_localTransform.pos.x += change;
+									_DMESSAGE("Updating X translation for %s by %f to %f", weapname, change, weap->m_localTransform.pos.x);
+									break;
+								case y:
+									weap->m_localTransform.pos.y += change;
+									_DMESSAGE("Updating Y translation for %s by %f to %f", weapname, change, weap->m_localTransform.pos.y);
+									break;
+								case z:
+									weap->m_localTransform.pos.z += change;
+									_DMESSAGE("Updating Z translation for %s by %f to %f", weapname, change, weap->m_localTransform.pos.z);
+								}
 							}
 						}
 						//_hasLetGoRepositionButton is always one frame after _repositionButtonHolding
@@ -2779,14 +2803,30 @@ namespace F4VRBody
 							writeOffsetJson();
 						}else if (_hasLetGoRepositionButton && _pressLength > 0 && _pressLength > c_holdDelay) {
 							auto change = vec3_len(_endFingerBonePos) > vec3_len(_startFingerBonePos) ? vec3_len(_endFingerBonePos - _startFingerBonePos) : -vec3_len(_endFingerBonePos - _startFingerBonePos);
-							weap->m_localTransform.pos.x += change;
-							_MESSAGE("Saving position translation for %s from %f -> %f", weapname, _customTransform.pos.x, weap->m_localTransform.pos.x);
-							_customTransform.pos.x = weap->m_localTransform.pos.x;
+							switch (_repositionMode) {
+							case x:
+								weap->m_localTransform.pos.x += change;
+								_MESSAGE("Saving X position translation for %s from %f -> %f: powerArmor: %d", weapname, _customTransform.pos.x, weap->m_localTransform.pos.x, _inPowerArmor);
+								_customTransform.pos.x = weap->m_localTransform.pos.x;
+								break;
+							case y:
+								weap->m_localTransform.pos.y += change;
+								_MESSAGE("Saving Y position translation for %s from %f -> %f: powerArmor: %d", weapname, _customTransform.pos.y, weap->m_localTransform.pos.y, _inPowerArmor);
+								_customTransform.pos.y = weap->m_localTransform.pos.y;
+								break;
+							case z:
+								weap->m_localTransform.pos.z += change;
+								_MESSAGE("Saving Z position translation for %s from %f -> %f: powerArmor: %d", weapname, _customTransform.pos.z, weap->m_localTransform.pos.z, _inPowerArmor);
+								_customTransform.pos.z = weap->m_localTransform.pos.z;
+							}
 							_hasLetGoRepositionButton = false;
 							_useCustomWeaponOffset = true;
 							g_weaponOffsets->addOffset(weapname, _customTransform, _inPowerArmor);
 							writeOffsetJson();
 						}
+						else if (_useCustomWeaponOffset 
+							&& !(_hasLetGoRepositionButton || _repositionButtonHolding) // do not allow defaults when handling reposition
+							&& reg & vr::ButtonMaskFromId((vr::EVRButtonId)c_defaultPositionButtonID)) {
 							_MESSAGE("Resetting grip to defaults for %s: powerArmor: %d", weapname, _inPowerArmor);
 							_useCustomWeaponOffset = false;
 							g_weaponOffsets->deleteOffset(weapname, _inPowerArmor);
